requirements.txt
python-dotenv==1.0.1
numpy==1.26.4
ultralytics==8.3.94
torch==2.6.0
torchvision==0.21.0
opencv-python==4.11.0.86
pillow==11.1.0
transformers>=4.34.0
huggingface-hub>=0.16.4
timm>=1.0.15